Corrected the units error in photontable target_coordinates and the temporal vs cube out format. Several changes to drizzler.py...
DitherDescription significantly reduced and renamed to DrizzleParams. It calculates and consolidates the relevant info for the different Drizzler()s

The default behaviour is now to use the metadata ra and dec as opposed to the simbad lookup of the target name. Closes #43. Should the user want to use the simbad lookup they flip the simbad_lookup bool in DrizzleParams

StarRA and StarDec renamed to centerRA and centerDec since the user might not always want to center on a star

Few remaining instances of min_timestep -> max_timestep

Removed the no-longer supported SpectralDrizzer class and added a todo to reintroduce that functionality if we want it

Fixed a super bug closes #44

timestep is now a parameter in the out config file and is read by form in run_pipeline

Added a sainity check logger statement at the end of form